+++ admin/routes/information.py
@@ -24,9 +24,10 @@ def information_add():
 def information_delete(id):
     from run import db
     information=OurInformations.query.get(id)
-    db.session.delete(information)
-    db.session.commit()
-    return redirect('/admin/information')
+    if request.method=='POST':
+        db.session.delete(information)
+        db.session.commit()
+        return redirect('/admin/information')
 
 @admin_bp.route('/information/edit/<id>', methods=['GET','POST'])
 def information_edit(id):
