+++ admin/routes/portfoliocategory.py
@@ -22,9 +22,10 @@ def portfoliocategory_add():
 def portfoliocategory_delete(id):
     from run import db
     portfoliocategory=PortfolioCategory.query.get(id)
-    db.session.delete(portfoliocategory)
-    db.session.commit()
-    return redirect('/admin/portfoliocategory')
+    if request.method=='POST':
+        db.session.delete(portfoliocategory)
+        db.session.commit()
+        return redirect('/admin/portfoliocategory')
 
 @admin_bp.route('/portfoliocategory/edit/<id>', methods=['GET','POST'])
 def portfoliocategory_edit(id):
