+++ admin/routes/pricing.py
@@ -23,9 +23,10 @@ def pricing_add():
 def pricing_delete(id):
     from run import db
     pricing=Pricing.query.get(id)
-    db.session.delete(pricing)
-    db.session.commit()
-    return redirect('/admin/pricing')
+    if request.method=='POST':
+        db.session.delete(pricing)
+        db.session.commit()
+        return redirect('/admin/pricing')
 
 @admin_bp.route('/pricing/edit/<id>', methods=['GET','POST'])
 def pricing_edit(id):
