+++ admin/routes/pricingoption.py
@@ -25,9 +25,10 @@ def pricingoption_add():
 def pricingoption_delete(id):
     from run import db
     pricingoption=PricingOptions.query.get(id)
-    db.session.delete(pricingoption)
-    db.session.commit()
-    return redirect('/admin/pricingoption')
+    if request.method=='POST':
+        db.session.delete(pricingoption)
+        db.session.commit()
+        return redirect('/admin/pricingoption')
 
 @admin_bp.route('/pricingoption/edit/<id>', methods=['GET','POST'])
 def procongoptions_edit(id):
