+++ admin/routes/service.py
@@ -30,11 +30,12 @@ def service_add():
 def service_delete(id):
     from run import db
     service=Services.query.get(id)
-    filename=f"./static/uploads/{service.image}"
-    os.remove(filename)
-    db.session.delete(service)
-    db.session.commit()
-    return redirect('/admin/service')
+    if request.method=='POST':
+        filename=f"./static/uploads/{service.image}"
+        os.remove(filename)
+        db.session.delete(service)
+        db.session.commit()
+        return redirect('/admin/service')
 
 @admin_bp.route('/service/edit/<id>', methods=['GET','POST'])
 def service_edit(id):
