+++ admin/routes/teamsimage.py
@@ -35,8 +35,9 @@ def teamsimage_add():
 def teamsimage_delete(id):
     from run import db
     image=TeamImages.query.get(id)
-    filename=f"./static/uploads/{image.image}"
-    os.remove(filename)
-    db.session.delete(image)
-    db.session.commit()
-    return redirect('/admin/teamsimage')
+    if request.method=='POST':
+        filename=f"./static/uploads/{image.image}"
+        os.remove(filename)
+        db.session.delete(image)
+        db.session.commit()
+        return redirect('/admin/teamsimage')
